csi: Get heal info from server pods
