+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CSI: Fetch volfiles from brick processes in server using --volfile-server &
   remove storage options from mount flow.
 - Server: Disbale default client volfile options
+- CLI: Move healinfo to server pods, Trigger client-heal from csi pod.
 
 ## [0.9.0] - 2022-11-21
